Update to bindgen 0.61.0
Summary:
T133646286 is going to require a fix in bindgen. This diff updates bindgen from 0.60 to the most recent major version 0.61 so that as soon as a fix for the crash is available, we can have an easier time pulling it into fbsource.

The major difference between bindgen 0.60 and 0.61 is the bindgen library (for build scripts to use) and the bindgen command-line tool (which we use in buck genrule) are no longer distributed in the same package. The cli moved to a `bindgen-cli` package. This impacts the implementation of the `rust_bindgen_library` Buck macro in fbcode_macros.

Reviewed By: zertosh

Differential Revision: D40539114

fbshipit-source-id: 3ae3237dca0094c8cee864c15a739b101d1c715d