wal healer: implement healing strategy
Summary:
New healer reads a batch of entries from the WAL, each entry corresponds to a single multiplexed write operation. Thus a batch contains unique <blobstore_key, operation_key>` entries but can have more than one entry per `blobstore_key`.

For each blob the healer tries to read this blob from the storages specified in the multiplexed config and figures out those stores which are missing the blob. Next it tries to upload the blob to the missing blobstores and keeps track of those blobstores, where writes failed.

Processed queue entries are deleted, while entries for the blobs which were not properly healed are requeued with a new timestamp for another healing attempt.

In case a blob is missing everywhere the healer currently just writes a warning, but it should log it to scuba, so we could raise an alarm.

Reviewed By: yancouto

Differential Revision: D40252552

fbshipit-source-id: 490097d9ea5be388d111e8e1894b46bd928a75b1