config: add a few functions for exposing the user, local, and repo paths
Summary:
Adds two functions for getting the user and system configs based on the identity to the `configparser` crate, as well as a method for getting the config file for the non-backing repo in the repo object.

This will be used later in a command for showing these to the user.

Reviewed By: bolinfest

Differential Revision: D40662810

fbshipit-source-id: c035233f627fd6847a2d48825138bc4db6e95b45