dirstate: move dirstate writing out of watchmanfs
Summary:
In a later diff we'll want to write the dirstate from Rust. Right now
the only place that does this is in watchmanfs. Let's refactor that logic out to
the treestate's dirstate module so it can easily be reused.

I also added some additional logic to update the dirstate parents based on the
current treestate parents.

Reviewed By: jordanwebster

Differential Revision: D40576684

fbshipit-source-id: c6069697f819db1e148ec5b8453b367f2df3b6f9