checkout: remove n^2 behavior in pathauditor
Summary:
The pathauditor was validating every component of every path written at
checkout time, with no short-circuiting even if the path had been validated
before.  This diff changes it to audit in bottom-up order, and to short-circuit
as soon as it sees a path it has already audited.

In my profiling, this code was taking 10% of the checkout cpu before my
optimization.

Reviewed By: muirdm

Differential Revision: D40573908

fbshipit-source-id: c6888065befe2fe2a456ebe4f24c453b24dcdf57