split file content management from IOverlay
Summary:
This diff does a few things:

1) removes `File` operations from `IOverlay`
2) moves the core of the implementation of `FsOverlay` into `FileContentStore`. `FsOverlay` stores a raw pointer to a `FileContentStore` object and redirects most functions to that object.
3) Adds a `FileContentStore` to `Overlay`. This is used as the core of `FsOverlay` and is used to manage `BlobInode` functions.

The goal of this diff is to create a structure that would allow for Linux and macOS to use other `IOverlay` implementations while maintaining access to the `File` contents of the `Overlay` and the `InodeMetadataTable`. With this setup, a `FileContentStore` will always exist on non-Windows platforms, and `backingOverlay_` could either be a `FsOverlay` which basically wraps the `fileContentStore`, or an independent `[Buffered]TreeOverlay` object.

This diff technically allows for the usage of a `TreeOverlay` on non-Windows, but `OverlayChecker` isn't set up to handle different `IOverlay` implementations, so a `throw` is added in `getTreeOverlayType` to prevent usage as of now.

Reviewed By: chadaustin

Differential Revision: D40154911

fbshipit-source-id: 383253ab456a979ef481fb2c0547fc16a32ccab7