oss: add windows build script
Summary:
This is a minimal windows build script based on build_nupkg.py script. For now we are producing a self contained zip file. In the future we probably want to produce an actual installer.

Openssl is assumed to be installed someplace the openssl-sys crate can find. The two viable options are to install openssl using vcpkg, or to set the "vendored" feature for the openssl crate so it builds from its vendored source.

Reviewed By: bolinfest

Differential Revision: D40703552

fbshipit-source-id: 60e20b638bd05b9755740e5e71bf3e36db8548fd