Fallback to x86_64 if running on ARM64 and edenfs is x86_64
Summary:
# Problem

If you try to run `edenfs` daemon which has been compiled for x86_64 on ARM64, you will get an because there's an assumption in `daemon.py` that ARM64 slice will be always available.

That's not the case when developing EdenFS on ARM64 but building x86_64 binaries (the default).

# Solution

As per `man arch`, we can just add additional architectures which will be attempted to be loaded.

```
If more than one architecture is specified, the operating system will try each one in order, skipping an architecture
that is not supported on the current processor, or is unavailable in the universal binary.
```

So, on an ARM64 machines, always prefer to run native but allow fallback to x86_64 if there's no ARM64 arch slice.

Reviewed By: xavierd

Differential Revision: D40794761

fbshipit-source-id: 838d1ffca73ebfd6e6c2265689965d709630b91a