rename IOverlay to InodeCatalog
Summary: Having an `Overlay` object own a `IOverlay` object, as opposed to deriving from one, is a bit confusing, so this renames `IOverlay` to `InodeCatalog` as an attempt to clear up the class structure. `InodeCatalog` is meant to mean "InodeDatabase", but I find the word "database" to imply a heavyweight structure, which this class is not.

Reviewed By: chadaustin

Differential Revision: D40517958

fbshipit-source-id: a5873ab21a336da3c6823b91d7f51c4ae49a7f08