Adding queue to process batch of requests
Summary:
For short, we add two channels. The mpsc channel to put all requests in a common channel. Note that we pass the sender of the oneshot channel in the mpsc channel. Hence, the oneshot channel to be able to retrieve the request. Right, now we keep processing each request individually even though we wait the batch process.

Next, we will group similar requests together. More info about it in the upcoming diff.

Reviewed By: RajivTS

Differential Revision: D40754635

fbshipit-source-id: 09b36037f069dffb5fa4c009e998cac6dc38e63d