Optimize get_nb_commits check
Summary:
There's no need to query git to see how many commits are unsyced because we already know if the root was synced or not when we checked it during missing_for_commit.

This is mostly just for the sake of code cleanliness, but our logs suggest this step takes ~3ms, so hopefully it will shave off a bit of time as well. (Not sure how much of the 3ms is logging overhead vs the actual git call though.)

Unfortunately, there's a second use in gitimport_acc which is used in test logging, so I left it in place for now.

Reviewed By: mojsarn

Differential Revision: D40944431

fbshipit-source-id: 27ac958e960f3448f6ebf1a074f50b39ffe1cf82