inodes: add a isInodeLoadedOrRemembered helper
Summary:
During invalidation on Windows, we need to know if the inode is known by the
InodeMap. This is currently achieved by testing the loaded inode set and the
unloaded one, taking the InodeMap lock twice, and constructing an InodePtr to
destroy it immediately.

The InodePtr destruction is potentially deadlock inducing as that may require
taking the parent lock, which is already taken. To avoid this, the newly added
method merely test both set without constructing an InodePtr.

Reviewed By: genevievehelsel

Differential Revision: D40956445

fbshipit-source-id: 662c139634e9ce79afacabcff2aeb9a682fc33b2