wal: Maximum number of retries for blob not found
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

Currently, when the healer doesn't find a blob in any blobstore, it puts it back in the queue.

However, it the blob did fail to write on all blobstores, it will stay in the queue forever. This diff adds a limit to how many times an item can be put back in the queue (magic number 20).

This makes it necessary to change the SQL schema, but that's fine because we haven't started putting this in prod yet.

Differential Revision: D40715865

fbshipit-source-id: 55a6e4dd0c7bb6ae36ba6b2d724e7a2260367b56