wal: Add sharded db config to queue
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

This syncs configerator changes from D40677044 and updates our code so that WAL config now takes a sharded DB instead of unsharded. The code for dealing with sharded connections was already added on D40673033, this is just the config to enable it.

For now this is not enabled anywhere. The plan is to:
- Test it locally with locally built binaries
- Move a small-ish repo
- Move everything

This is once we have full feature parity, which is still not done (things like Scuba logging and scrub blobstores are missing)

Differential Revision: D40675512

fbshipit-source-id: 70f74a53dbf10eadc73ce9e210ec7719f866e30a