improve error message w/ invalid Windows std handles
Summary:
Improve errors so it is clearer hg was invoked with invalid stderr/stdin/stdout handle(s).

A further improvement may be to try to ignore the errors and continue operating in this case, but that raises questions (e.g. "is it a good idea to run if you can't output information to the user?"). Let's start with a better error for now.

Reviewed By: zzl0

Differential Revision: D40878799

fbshipit-source-id: a1135262d033407276114b8c5304c5e5aadc9fb7