cli: reconfigure the output to allow unicode errors
Summary:
Python on Windows has a weird behavior when redirecting stdout/stderr as it
sets the encoding to cp1252. Unfortunately that encoding is only a subset of
unicode and doesn't recognize some of the characters we're trying to print to
it. In particular, when displaying the backtrace during `eden doctor`, it would
throw:

  UnicodeEncodeError: 'charmap' codec can't encode character '\u2502' in position 89: character maps to <undefined>

Instead of that, let's replace the character with a `?` to get a better error.

Reviewed By: genevievehelsel

Differential Revision: D40997359

fbshipit-source-id: f9def7b0c663d5e8a9f5dd23c9aea0c5b24609fd