version: add setup_with_version.py wrapper script
Summary:
A wrapper script that is responsible for computing the correct version (or propagating through a version if it was set externally, e.g. GitHub workflow) and then passing it to the underlying setup.py script.

Updated Makefile, OSS Windows buildscript and build_nupky.py to call the new wrapper.

Reviewed By: bolinfest

Differential Revision: D40983642

fbshipit-source-id: e76033e48a7eaf4347da63eae273806e238f7cdb