repolock: always clean up legacy lock path
Summary:
The rust repolock still creates the legacy lock files (e.g. .hg/wlock). Previously in the python_and_rust transitional lock mode, rust would try to check if the Python code was "really" using the legacy paths to avoid deleting the legacy lock files out from under Python. Now that we are in rust_only mode, this logic is not necessary, and in fact can cause the .hg/wlock file to get stuck on Windows if an interrupted operation doesn't clean up .hg/wlock for whatever reason.

On Windows, the legacy locking is achieved using the "you can't rename file into an existing file" property, so the mere presence of the file means the file is "locked". So, if the wlock file failed to get cleaned up for whatever reason, Rust would see that it already existed and leave it alone, assuming Python would clean it up. But with Python locking disabled, that means no one would clean it up.

Reviewed By: sggutier

Differential Revision: D41005455

fbshipit-source-id: a83e92208f9f9a87a528e501a5f4a31ede948c5c