support make oss RUST_DEBUG=1
Summary:
`make oss` builds the Rust code in release mode by default.
This adds logic to our `Makefile` to check for a `RUST_DEBUG=1`
argument that, if set, adds the `--debug` flag to `build_rust_ext`.

This should be helpful when iterating on local development
and can also be used in smoke builds to get signal faster.

Reviewed By: zzl0

Differential Revision: D41063517

fbshipit-source-id: cd7b14495ba01ff01d9889f45c31270573fc24c4