unify wrapper code for GitHub CLI
Summary:
As noted by the TODO in `eden/scm/ghstack/github_gh_cli.py`,
this code was initially copypasta'd because I thought I was
going to have to maintain both synchronous and asynchronous
versions. In the end, I was able to use the async version in ghstack
by wrapping it with `asyncio.get_event_loop()` and `loop.run_until_complete()`.

While iterating on the ghstack version, I ended up making some
improvements to (1) support the GitHub REST endpoints as well
as (2) boolean arguments, so now the
`eden/scm/edenscm/ext/github/gh_submit.py` code can get the
benefits of this, as well.

Reviewed By: jordanwebster

Differential Revision: D41053270

fbshipit-source-id: e36032088df02bdd68693d03276b3571cf6cfcc7