remove escaped braces now that it is no longer an f-string
Summary:
Before landing D41062213 (https://github.com/facebookexperimental/eden/commit/2736052d39852f48409d8e2404d67d113ecc5482), I got a lint warning that I
had an unnecessary use of an f-string because there
weren't any expressions evaluated as part of the f-string.
I removed the `f` and just hit Land without re-running
`gen_workflows.py`.

In this case, removing the `f` changed the behavior of
the code because the string originally included double `{{`
to escape it, which is necessary as an f-string, but is
interpreted literally in an ordinary string literal.

This diff removes the double `{{` so that `gen_workflows.py`
produces the desired output again.

Reviewed By: sggutier

Differential Revision: D41105429

fbshipit-source-id: f4bc023cd42a7f265325fd9262e7c602e900bb99