further fix config dir logic
Summary:
In some cases, automation executes a command outside of a mount specified with the --mount flag (i.e. cwd is not a child directory of the repo we wish to run the command with). In those cases, errors will occur if the config dir we wish to use is not the default config dir. Therefore, we must add the following logic:

1. if a subcommand takes a --mount argument, use that to infer which config dir to use
2. if no --mount option is supplied but a --config-dir argument is supplied, use the supplied config dir.
3. If neither a --mount or --config-dir argument is supplied, infer the config dir based on the cwd.
4. If the cwd is not part of an eden repo, fall back to the default config dir location.

Reviewed By: fanzeyi

Differential Revision: D41050189

fbshipit-source-id: 68062a88d9ca9f4e9c96a8539a368745d49d3ad3