checkout: fallback if Rust commit resolution fails
Summary:
Rust native checkout doesn't support all revsets yet, so let's just
fallback to Python if the commit resolution fails.

This diff also fixes a small issue with revset resolution where 'hg up tip' in
an empty repo causes the revset resolution to return b"" and tries to parse that
as a HgId and fails. Let's just have it return null_id in that case.

Reviewed By: quark-zju

Differential Revision: D40735127

fbshipit-source-id: d5f1f6c42f3eab0ce02eef9f04eacf03f656e0ac