wal: Return shard when inserting entries
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

Running a `SqlBlobstoreWal::log` query now returns entries instead of simply consuming them. The entries have their `shard_id` field set. This will be used on D40905438 to be able to delete those entries in an optimisation scenario.

Since in this case we have a `shard_id` but not a `id`, I needed to refactor the `ReadInfo` struct to have two nullable fields.

**Why doesn't it also return `id`?** It *could* be doable, but the [docs](https://dev.mysql.com/doc/refman/8.0/en/information-functions.html#function_last-insert-id) for that are quite complicated, and it becomes impossible without an extra query if we start using insert_or_ignore.

**Possible next step**: I'm considering making all of this strongly typed, though that will mean we will have multiple "entry types".
Something like:
- `BlobstoreWalEntry`: All fields, no id or shard id. Used to call `log/log_many` only.
- `BlobstoreWalEntryRead`: Returned by `SqlBlobstoreWal::read`, always has `id` and `shard_id` can be used on  `SqlBlobstoreWal::delete`.
- `BlobstoreWalEntryWrite`: Returned by  `SqlBlobstoreWal::log/log_many`, always has `shard_id` but not `id`. (Will be usable in `SqlBlobstoreWal::delete_key`)

This option is more technically correct, avoids usage mistakes on compile time, but has a bit more boilerplate and may be a bit clumsy when you want to use one object on another place. Seems more rusty, though, integers are like that in Rust for example, you always get explicit errors and need to be very careful with conversions.
What is the opinion on this?

Reviewed By: mitrandir77

Differential Revision: D40905439

fbshipit-source-id: be09e5d875ceaa8fd1a329e25ffe34d6108b3994