wal: Add scuba logging for multiplex
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

This reuses the functions extracted on D40724724 to use the same logic as the current multiplex to do scuba logging for the new multiplex.

The only notable difference to the old blobstore here is that `record_queue_stats` is only called once and not once per inner blobstore (per design of WAL), so we don't have the blobstore inner id when calling it.

Differential Revision: D40724723

fbshipit-source-id: e4f9904d420a3e928ab85b9db0f401533780c599