multiplexed: Extract scrub repair to a separate method
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z
Scrub blobstore is a blobstore that tries to repair data across the inner storages as it traverses them. It is used by the walker.

This extracts the logic to repair some of the inner blobstores to another function, so we can reuse it in the WAL scrub blobstore later on.

It shouldn't change any behaviour.

Reviewed By: markbt

Differential Revision: D40810430

fbshipit-source-id: 927e48a3e3d4fb2c0ce83488128576e11b12fa7b