multiplex: Make errors separate main errors and write mostly
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z
Scrub blobstore is a blobstore that tries to repair data across the inner storages as it traverses them. It is used by the walker.

This splits errors from main blobstores from errors from write mostly blobstores in some of our error types. This is to allow finer grained error handling. The write_mostly errors are not as important as the main errors.

Reviewed By: RajivTS

Differential Revision: D40858674

fbshipit-source-id: ab109d2e166f4ad3a2ef08ea91bafc1b0d9daa82