wal: Add scrub tests
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z
Scrub blobstore is a blobstore that tries to repair data across the inner storages as it traverses them. It is used by the walker.

This diff adds unit tests for the WAL scrub blobstore. They are based on the current scrub blobstore tests.

It doesn't change behaviour of the scrub blobstore itself, but I added some comments or context in some places where the logic was surprising to me.

The most tricky part here is around failures of inner blobstores. Since scrub is trying to repair stuff, we can't really repair things that are broken. So the scrub blobstore can either (1) fail if an underlying blobstore is down, or (2) try to repair as much as possible of what's not failing. I'm trying to go with (2).

Here are the behaviours that surprised me, or that work different on the new WAL scrub blobstore:
- The current one looks at the sync queue to decide if something is "in the process of healing". WAL doesn't do that, scrub shouldn't need to know about healer.
  - This means if some blobstores fail and others return None, we don't know if the blob exists or not. Currently, we always fail on this case, but I plan to change that on D40858675 so that we take quorum into account.
- Current behaviour: If one blobstore succeeded at returning `Some(value)`, we repair the blobstores that return `None`, but completely ignore all blobstore that failed, and consider the whole operation a success if the non-errored blobstores repair successfully. Should this be different?

Reviewed By: mitrandir77

Differential Revision: D40857369

fbshipit-source-id: 1fd7688eafd7bbc26eec61aeacfcfc091670cc65