multiplexed: Extract behaviour of record_queue_stats
Summary:
This is a refactor only change. I plan to reuse this function, so let's extract it to a different file and simplify how to use it.

Using a generic argument on `record_put` allow us to use it with more types of results, which will also be useful when reusing this.

Differential Revision: D40724722

fbshipit-source-id: 4211d85876fd2ad696466150e778807b58ce7f2b