wal: Update scuba config, don't use multiplex scuba yet
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

This sync configerator changes from D40754273, and updates our internal configs to load two types of scuba tables. But it continues doing the same as before, only logging to scuba on internal blobstore queries.

I also renamed `scuba` to `inner_blobstores_scuba` to make it clear, because that's confusing on the current multiplexed blobstore.

Note: the thrift compat error is fine to bypass because we're not using this thrift schema anywhere yet.

Differential Revision: D40720597

fbshipit-source-id: fceb9d72b20e298d52744dfd79798c36f34b07da