wal: Use read quorum to see if a blob is gone when blobstores fail
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z
Scrub blobstore is a blobstore that tries to repair data across the inner storages as it traverses them. It is used by the walker.

Currently, on WAL scrub we fail if **any** blobstore fails and the rest return `None`. But that means a single broken blobstore stops the scrub blobstore from progressing, which is bad. This diff makes it so we trust read quorum in this case, so a single broken blobstore shouldn't be a problem and we should repair "as much as possible".

Reviewed By: RajivTS

Differential Revision: D40858675

fbshipit-source-id: 1dd7370426b2750e732cf165845495f3b3424dbd