wal: Delete key from WAL queue when write is fully successful
Summary:
**Context**: We want to roll out the new WAL multiplexed blobstore and need final tweaks before doing that. See https://fb.quip.com/vyTWAgGZSA2Z

This implements the suggestion given here: https://fb.quip.com/vyTWAgGZSA2Z#temp:C:FUA14bc69fc6a824c279926a39ff

If absolutely all writes succeeded, we can remove the item from the queue, so the healer doesn't need to do the no-op of consuming it.

This should actually considerably lower the load on the healer, because the vast majority of cases succeed on all writes, so the queue would almost always be empty. It is a very valuable

Reviewed By: mitrandir77

Differential Revision: D40905438

fbshipit-source-id: f5ab10612b33e42b15033d6ec803ee4ed02794db