Removing fallback from land service
Summary:
To finish the land service migration, we will remove the land service pushrebase tunable.

Before land it, we will change the land service tunable flow to 95%, to see if the scs server is getting requests. Once we are sure that everthing is working fine, we will remove the tunable related to it.

Reviewed By: RajivTS

Differential Revision: D41026835

fbshipit-source-id: a1c4551b04ffb4940cc1fe83e453982fe707e83e