trace: fix with UNC AbsolutePath
Summary:
UNC paths broke the tracing code due to taking a user provided path as an input
and building an AbsolutePath from it. Since that path is unlikely to be a UNC
one, the AbsolutePath constructor throws and `eden trace` fails.

Even with this, starting `eden trace fs` makes any subsequent IO fail due to
EdenFS trying to construct an AbsolutePath from a NT Device path (\Device)
which isn't recognized.

Reviewed By: chadaustin

Differential Revision: D41170993

fbshipit-source-id: cba61f6e00942a18b36f9e801fdaeb565b2d91d1