prjfs: fix eden trace fs
Summary:
The EdenFS code was building an AbsolutePath from a path provided by
ProjectedFS, which unfortunately is neither a device letter path (C:), or a UNC
path (\\?\), but a NT path (\Device). The latter is unfortunately not
recognized by the AbsolutePath code and thus an exception is raised, failing
the IO.

Instead of building an AbsolutePath, let's manually find the basename and build
a string out of it.

Reviewed By: kmancini

Differential Revision: D41171761

fbshipit-source-id: 2133006a07b1cc27111b33c22efca6a461f9cdb1