Move pushredirection tunable check to pushredirection method
Summary:
All (the single) usage of this function checked the killswitch and used `None` if it was enabled. Let's move the killswitch check to the function to make sure it is hard to misuse.

I also changed the error type as it is always a service error, and this makes it more reusable in the future.

Reviewed By: RajivTS

Differential Revision: D41122814

fbshipit-source-id: a4198293f47d2cb669ee84194623844169ee282a