Create push_redirector in RepoContext and not in SCS
Summary:
This was a suggestion by mitrandir77

Instead of ad-hoc creating a push redirector for just the land_stack request, let's always create the redirector on the RepoContext, so that it is always available to all Mononoke API methods.

This makes code simpler, and has the nice side effect of making land_stack work with pushredirecton on EdenAPI for free! (Or automatically for anything else that uses mononoke_api)

Reviewed By: mitrandir77

Differential Revision: D41184507

fbshipit-source-id: c04cc9abce812c5dcf5bc48168923e6108e47d7b