Add UID to edenfs mount sanity checks
Summary:
We are trying to enable rootsquash on our netapp volume mounts (eg. userA sudos to root and does ls on some nfs mount -> the nfs server maps that root uid to some facade uid)

and we see that eden clone fails either the access, or stat checks. I tested an `access` performing binary and stat, and both let my personal user access the mount that fails in clone, so I'd like to add this extra int in the debug to clear it up. It seems as though sudo (and hence the facade account) is invoked since the facade account does show these errors.

Reviewed By: xavierd

Differential Revision: D40764092

fbshipit-source-id: 8df782b2f4c10cfa51186074753c1b0f75275409