Use u128 for hash key on SQL cache
Summary:
**Context**: We're adding caching to mononoke SQL queries. See D40553048

For the keys of the cache, we use hashes (we don't need to convert back to the original request from the key name).

This diff changes from using u64 to u128 as the hash size. This should prevent any hash collision issues that may arise. It also starts taking the query strings into account in the key, so that any change to the query will invalidate the cache, which is desired to avoid possible problems with bad cache.

Before, it would take about a billion cache keys being added before we started having a good chance of collision, which is already pretty big. With 128 bits, it only has a reasonable chance of collision with ~10^18 items, which we'll never ever have. (See [birthday paradox](https://en.wikipedia.org/wiki/Birthday_problem))

Reviewed By: RajivTS

Differential Revision: D40559019

fbshipit-source-id: c5b1f5aff854f10b6de77a639cd427e2ba11d6d0