Rename retryable queries to mononoke queries
Summary:
On the rest of this stack, I plan to also add caching for queries, not only retrying. So let's name this more appropriately.

This macro will be our generic layer on top of SQL that we can use to add generic features for all our SQL queries, in some way similar to how we add a bunch of stuff on top of Manifold/blobstores.

Reviewed By: RajivTS

Differential Revision:
D40553047

LaMa Project: L1122763

fbshipit-source-id: 96abc7f477b82db9018fbf7b512b020aeb0e282d