Using cacheable on a single query
Summary:
**Context**: We're adding caching to mononoke SQL queries. See D40553048

This diff adds the `cacheable` keyword to a query (select bubble, the one we saw problems with last week), and makes sure that it works. To do that, we needed to:
- Add `SqlQueryConfig` to one of our repo objects.
- Implement the creation of `SqlQueryConfig` on the repo factories.
- Pipe it through the `RepoEphemeralStore` methods in order to get it to the appropriate query.

Caching can be enabled on a read query by:
- Adding "cacheable" keyword to your query.
- Make sure all parameters (input) to the query implement the Hash trait.
- Making sure the return values (output) implement Serialize, Deserialize, and Abomonation.

Notice that this diff does not add any actual usage of cacheable queries, nor does it add a way to actually build a `SqlQueryConfig` object when running Mononoke. That will come in the next diff.

This involved some changes, like making caching objects optional, as we want to disable it in intergration tests and allow `TestRepoFactory` to build our object.

Reviewed By: mitrandir77

Differential Revision:
D40553046

LaMa Project: L1122763

fbshipit-source-id: b4dc2f646c5bf945e44ddc894e980a481b6f3b15