repolock: add new RepoLocker abstraction
Summary:
Instead of using raw lock functions, let's introduce a RepoLocker
abstraction that can be stored on the Repo and WorkingCopy. It is responsible
for managing the various repo locks and validating that they are taken in the
appropriate order.

Implements lock_store while we're at it since it's just a few more lines now.

Reviewed By: muirdm

Differential Revision: D40773686

fbshipit-source-id: 2789bab670544be4cab4ced55fc9a269b3d217fb