Remove unnecessary dependencies on `cmdlib`
Summary:
As crates were migrated from `cmdlib` to `mononoke_app`, I noticed some
of them kept the old dependency for only one struct (`AliveService`) which is also exposed
through `mononoke_app`.

Find those places and use the `mononoke_app` version to remove the dependency.

Note that for `segmented_changelog_tailer`, I couldn't remove the dependency in
this diff as it also uses `cmdlib::helpers::csid_resolve` for which there isn't
currently an existing equivalent.

Crates which lost the dependency on `cmdlib`:
* `//eden/mononoke:aliasverify`
* `//eden/mononoke:blobstore_healer`
* `//eden/mononoke/facebook/derived_data_service:derivation_worker`
* `//eden/mononoke/microwave:builder`
* `//eden/mononoke/repo_import:repo_import`

Crates which kept the dependency, but not for `AliveService`
* `//eden/mononoke:segmented_changelog_tailer`

Reviewed By: RajivTS

Differential Revision: D41184176

fbshipit-source-id: 58fc4fa18d029ba681ae50ae321066632dc1bf38