github: prefetch PR status using smartset.prefetch
Summary:
Currently, `sl ssl` is very slow if you have a tall stack, this diff is
to optimize that with prefetch, this can reduce the time from **22s -> 5s**
from my test with 15 PRs.

Currently, the GitHub GraphQL query does not support multiple PRs, the
next step is to make the prefetch in parallel.

Reviewed By: bolinfest

Differential Revision: D41225480

fbshipit-source-id: 96ba55737960ff1fe83507b0fffff9b87885f5f3