commands: integrate legacy aliases concept into @command
Summary: Add a "legacyaliases" keyword to the command decorator that can be used instead of the command_name helper. This is more ergonomic (don't need to import command_name), and it is nice to not have an extra level of indirection for generating command name. Also, we are back to only having a single way to specify aliases.

Reviewed By: bolinfest

Differential Revision: D41222449

fbshipit-source-id: 5151d9474ab29dfae09acac0c862e59e1a31946a