replace use of trait_alias crate with unstable trait_alias feature
Summary:
We already use `#![feature(trait_alias)]` in a bunch of places, so we might as
well use it instead of the `trait_alias` crate elsewhere.

There is one place this doesn't work, as we auto-impl the trait for `Box` and `Arc`.
In that case, use the `trait-set` crate instead.

Differential Revision: D41207875

fbshipit-source-id: 6de852b1e9fbf3d7542954d62ea27517ea21130b