add a mechanism for cross-compiling for macOS
Summary:
This diff allows cross-compiling from `aarch64-apple-darwin` to `x86_64-apple-darwin` and viceversa.

There's now an additional option in the Makefile for setting the `RUST_TARGET` variable, which sets the target for Rust. The target for the C parts of the codebase (e.g. `chg`) still needs to be configured via the `CFLAGS` environment variable, however.

Another caveat is that it's still necessary to provide native OpenSSL and Python libraries for the target architecture. This is somewhat made easy with the addition of the `./packaging/mac/prepare_environment.py` script, which makes it easy to download Homebrew bottles. One can just download the desired bottle containing the necessary libraries. For an example of how to do this, see the test plan below.

Reviewed By: muirdm, bolinfest

Differential Revision: D41188815

fbshipit-source-id: 692602856821be09d2e3f9786dda9a60093d394b