goto: update docstring
Summary:
- minimize aliases
- tweak wording
- normalize terminology

Note that I ran into really annoying problems due to the existence of this command in Rust and Python. The core problem is that commands are keyed in the command table by the concatenation of the command name and all the aliases. So, Rust and Python have to be exactly in sync, which is extra annoying when legacy aliases are involved. Anyway, I hacked around it for now, but I think the command table should be keyed by just the primary command name and aliases should be data within the table, and perhaps an additional index into the table.

Reviewed By: bolinfest

Differential Revision: D41258862

fbshipit-source-id: b9b3213fa03dc46f4addd5fd8f66ae4da35f3c61