unbundle: log new commits via repo_update_logger
Summary:
Log new draft or public commits created through unbundle via the new repo_update_logger for commits.

There is a subtle behaviour correction here: previously commits that were pushed via a bare push with no associated bookmark would've been treated as public and logged to the public scribe category.  However, this is incorrect: without a public bookmark, commits are draft, so these should be logged as new draft commits.  In practice we don't ever do this in production, but for consistency the code now follows this behaviour.

Reviewed By: mitrandir77

Differential Revision: D40600539

fbshipit-source-id: 7720f305023f50a7e147b0faa081ed862da0d93b