mononoke_api_hg: log new commits via repo_update_logger
Summary:
Log new draft commits created via `mononoke_api_hg` (which are commit cloud commits uploaded via EdenAPI) via the new repo_update_logger for commits.

Previously we only logged to scribe, and the scribe category was stored as part of the `CreateChangeset` object.  However, this was only used for this code path, and was set to `None` elsewhere.  Since we are now potentially logging to other locations, extract this out of the `CreateChangeset` code and call it directly.

We do still need to clean up creation of changesets, but that will come later.

Reviewed By: yancouto

Differential Revision: D40600541

fbshipit-source-id: 5ae5806eea0fe610f44b92d0ef61b937e735075b