init: fix init --git
Summary:
When using `sl`, init --git was failing with a `Invalid config item:
'remotefilelog.reponame' (reponame is not set)` error. This was caused by it
trying to create a treestate before the `git` requirement had been added. This
meant the treestate tried to create a file store which tried to instantiate
eden_api, which tried to read the reponame.

This seems to only have happened in arm64 builds. So far it seems to be because:

- Normally fsmonitor tries to load the dirstate, but fails because of the
  missing git requirement. fsmonitor has special logic that eats all exceptions,
  and instead just silently fallsback to not using fsmonitor.
- During this logic, it actually manages to write some partial dirstate info.
  The presence of this info actually causes later code to avoid trying to write
  a dirstate, and thus avoids the 'loading a dirstate before git requirment has
  been written' issue.
- But, in an arm64 build, fsmonitor fails to load because of a syscall failure
  during watchmanclient::getcanonicalpath(name). On arm64 it gives errno 14 for
  some reason. This causes fsmonitor to simply not load, and all that logic I
  mentioned earlier that does stuff and eats errors, doesn't get loaded at all.

The fix is to not load the treestate until the git store requirement has been
written.

Note, there is still the issue of watchman not loading in arm64 builds. But that
will be fixed later.

Reviewed By: sggutier

Differential Revision: D41288356

fbshipit-source-id: 368826ab2a40bf08db27078cd073baf2e1d2da1f