dag/render: include parent/ancestor information on other edge types
Summary:
Currently, the difference between parent and ancestor edges is only shown for the vertical component of a graph renderer cell.  This is incomplete: the horizontal or corner components may also be for either a parent or ancestor edge.

Expand the data model to include this information for all edges.  The ascii and box drawing renderers' outputs remain unchanged, but the ascii-large renderer is improved to show the difference between the two.  This makes the "ancestor" lines more clearly dotted.

Reviewed By: quark-zju

Differential Revision: D41286950

fbshipit-source-id: cb88c913926f10b2f971ee027d99599e13a012ab